requirements.txt
# # Core NLP & ML
# nltk>=3.8.0
# spacy>=3.5.0
# networkx>=3.0
# scikit-learn>=1.3.0

# # HTTP & Data
# requests>=2.31.0
# numpy>=1.24.0
# pandas>=2.0.0

# # LangChain / LangGraph (optional but recommended)
# langchain>=0.1.0
# langgraph>=0.0.1

# # Pathway (optional for streaming)
# pathway[xpack-llm]>=0.0

# # Config
# python-dotenv>=1.0.0


# Core NLP
spacy==3.7.5
nltk==3.9.2

# ML / Data
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.2
networkx==3.3
sentence-transformers==2.7.0
transformers==4.41.2

# RAG / LLM frameworks
langchain==0.1.20
langgraph==0.0.48
pathway[xpack-llm]==0.post1

# Utilities
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.67.1
